用户管理接口完善优化

+++ src/main/java/com/mbx/usercenter/controller/UserController.java
@@ -1,6 +1,10 @@
 package com.mbx.usercenter.controller;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.mbx.usercenter.common.BaseResponse;
+import com.mbx.usercenter.common.ErrorCode;
+import com.mbx.usercenter.common.ResultUtils;
+import com.mbx.usercenter.exception.BusinessException;
 import com.mbx.usercenter.model.domain.User;
 import com.mbx.usercenter.model.request.UserLoginRequest;
 import com.mbx.usercenter.model.request.UserRegisterRequest;
@@ -10,8 +14,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
+
+import static com.mbx.usercenter.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * 用户接口
@@ -23,11 +28,8 @@ import java.util.List;
 @RequestMapping("/user")
 public class UserController {
 
-
     @Resource
     private UserService userService;
-
-
 
     /**
      * 用户注册
@@ -35,17 +37,18 @@ public class UserController {
      * @return 新用户id
      */
     @PostMapping("/register")
-    public Long register(@RequestBody UserRegisterRequest userRegisterRequest) {
+    public BaseResponse<Long> register(@RequestBody UserRegisterRequest userRegisterRequest) {
         if(userRegisterRequest == null){
-            return null ;
+            throw new BusinessException(ErrorCode.PARAMS_ERROR);
         }
         String userAccount = userRegisterRequest.getUserAccount();
         String userPassword = userRegisterRequest.getUserPassword();
         String checkPassword = userRegisterRequest.getCheckPassword();
         if (StringUtils.isAnyBlank(userAccount, userPassword, checkPassword)){
-            return null;
+            throw new BusinessException(ErrorCode.PARAMS_ERROR, "账号、密码或确认密码不能为空");
         }
-        return userService.userRegister(userAccount, userPassword, checkPassword);
+        Long result = userService.userRegister(userAccount, userPassword, checkPassword);
+        return ResultUtils.success(result);
     }
 
     /**
@@ -54,18 +57,18 @@ public class UserController {
      * @return 脱敏后的用户信息
      */
     @PostMapping("/login")
-    public User login(@RequestBody UserLoginRequest userLoginRequest, HttpServletRequest request) {
+    public BaseResponse<User> login(@RequestBody UserLoginRequest userLoginRequest, HttpServletRequest request) {
         if(userLoginRequest == null){
-            return null ;
+            throw new BusinessException(ErrorCode.PARAMS_ERROR);
         }
         String userAccount = userLoginRequest.getUserAccount();
         String userPassword = userLoginRequest.getUserPassword();
         if (StringUtils.isAnyBlank(userAccount, userPassword)){
-            return null;
+            throw new BusinessException(ErrorCode.PARAMS_ERROR, "账号密码不能为空");
         }
-        return userService.userLogin(userAccount, userPassword, request);
+        User user = userService.userLogin(userAccount, userPassword, request);
+        return ResultUtils.success(user);
     }
-
 
     /**
      * 根据用户名查询用户
@@ -73,14 +76,16 @@ public class UserController {
      * @return 用户列表
      */
     @GetMapping("/getUserByUsername")
-    public List<User> getUserByUsername(String username, HttpServletRequest request) {
+    public BaseResponse<List<User>> getUserByUsername(String username, HttpServletRequest request) {
         //判断是否管理员用户
         if (!userService.isAdmin(request)) {
-            return new ArrayList<>();
+//            return ResultUtils.error(401, "无权限访问");
+            throw new BusinessException(ErrorCode.FORBIDDEN_ERROR);
         }
         List<User> userList = userService.list(new QueryWrapper<>(new User()).like("username", username));
         //用户数据脱敏
-        return userList.stream().map(user -> userService.userMasking(user)).toList();
+        List<User> maskedUserList = userList.stream().map(user -> userService.userMasking(user)).toList();
+        return ResultUtils.success(maskedUserList);
     }
 
     /**
@@ -89,21 +94,53 @@ public class UserController {
      * @return 是否删除成功
      */
     @PostMapping("/deleteUser")
-    public boolean deleteUser(@RequestBody Long id, HttpServletRequest request) {
-
-        if(id <= 0 ){
-            return false;
+    public BaseResponse<Boolean> deleteUser(@RequestBody Long id, HttpServletRequest request) {
+        if(id == null || id <= 0 ){
+//            return ResultUtils.error(400, "用户ID不合法");
+            throw new BusinessException(ErrorCode.PARAMS_ERROR, "用户ID不合法");
         }
 
         //判断是否管理员用户
         if (!userService.isAdmin(request)) {
-            return false;
+//            return ResultUtils.error(401, "无权限删除用户");
+            throw new BusinessException(ErrorCode.FORBIDDEN_ERROR);
         }
-        return userService.removeById(id);
+        boolean result = userService.removeById(id);
+        return ResultUtils.success(result);
     }
 
+    /**
+     * 获取当前用户
+     * @param request 请求
+     * @return 当前用户
+     */
+    @GetMapping("/current")
+    public BaseResponse<User> getCurrentUser(HttpServletRequest request) {
+        if (request == null) {
+//            return ResultUtils.error(400, "请求为空");
+            throw new BusinessException(ErrorCode.PARAMS_ERROR);
+        }
+        User user = (User) request.getSession().getAttribute(USER_LOGIN_STATE);
+        if (user == null) {
+//            return ResultUtils.error(20001, "未登录");
+            throw new BusinessException(ErrorCode.OPERATION_ERROR,"未登录");
+        }
+        long userId = user.getId();
+        user = userService.getById(userId);
+        return ResultUtils.success(userService.userMasking(user));
+    }
 
-
-
-
+    /**
+     * 用户注销
+     * @param request 请求
+     * @return 是否注销成功
+     */
+    @PostMapping("/logout")
+    public BaseResponse<Boolean> logout(HttpServletRequest request) {
+        if (request == null) {
+            throw new BusinessException(ErrorCode.PARAMS_ERROR);
+        }
+        userService.userLoginOut(request);
+        return ResultUtils.success(true);
+    }
 }
